requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
class-registry==2.1.2
config==0.5.0.post0
cycler==0.10.0
decouple==0.0.7
Django==3.1.4
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
filters==1.3.2
filters-django==1.0.5
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.0.2
numpy==1.19.4
Pillow==8.0.1
protobuf==3.14.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
